+++ v0.1/update_ip/src/domain_services/mod.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
-use crate::type_flyweight::{Config, DomainResult, UpdateIpResults};
+use crate::type_flyweight::config::Config;
+use crate::type_flyweight::results::{DomainResult, UpdateIpResults};
 
 mod dyndns2;
 mod cloudflare;
@@ -9,8 +10,9 @@ pub async fn update_domains(
     config: &Config,
     results: &UpdateIpResults,
 ) -> HashMap<String, DomainResult> {
-    // add more services here
     let mut domain_results = HashMap::<String, DomainResult>::new();
+    
+    // add more services here
     domain_results = dyndns2::update_domains(domain_results, results, config).await;
     domain_results = cloudflare::update_domains(domain_results, results, config).await;
