+++ v0.1/update_ip/src/ip_services/address_as_body.rs
@@ -1,7 +1,7 @@
 use std::net;
 
 use crate::requests;
-use crate::type_flyweight::IpServiceResult;
+use crate::type_flyweight::results::IpServiceResult;
 
 // request with empty body returns response body with IP Address
 pub async fn request_address_as_response_body(
