+++ v0.1/update_ip/src/type_flyweight/dyndns2.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct Dyndns2 {
+    pub service_uri: String,
+    pub hostname: String,
+    pub username: String,
+    pub password: String,
+}
